+++ src/main/java/com/esiitech/trombinoscope_api/service/PromotionService.java
@@ -1,44 +1,59 @@
 package com.esiitech.trombinoscope_api.service;
 
-
 import com.esiitech.trombinoscope_api.Entity.Promotion;
+import com.esiitech.trombinoscope_api.Exception.PromotionNotFoundException;
 import com.esiitech.trombinoscope_api.repository.PromotionRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
-
 
 @Service
+@RequiredArgsConstructor
 public class PromotionService {
 
     private final PromotionRepository promotionRepository;
-
-    public PromotionService(PromotionRepository promotionRepository) {
-        this.promotionRepository = promotionRepository;
-    }
 
     public List<Promotion> getAllPromotions() {
         return promotionRepository.findAll();
     }
 
-    public Optional<Promotion> getPromotionById(Long id) {
-        return promotionRepository.findById(id);
+    public Promotion getPromotionById(Long id) {
+        return promotionRepository.findById(id)
+                .orElseThrow(() -> new PromotionNotFoundException("Promotion avec ID " + id + " non trouvée"));
     }
 
     public Promotion createPromotion(Promotion promotion) {
+        if (promotionRepository.existsByAnnee(promotion.getAnnee())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Une promotion avec cette année existe déjà");
+        }
         return promotionRepository.save(promotion);
     }
 
     public Promotion updatePromotion(Long id, Promotion promotionDetails) {
-        return promotionRepository.findById(id).map(promotion -> {
-            promotion.setAnnee(promotionDetails.getAnnee());
-            return promotionRepository.save(promotion);
-        }).orElseThrow(() -> new RuntimeException("Promotion non trouvée"));
+        return promotionRepository.findById(id)
+                .map(promotion -> {
+                    if (promotion.getAnnee() != promotionDetails.getAnnee()
+                            && promotionRepository.existsByAnnee(promotionDetails.getAnnee())) {
+                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Une autre promotion avec cette année existe déjà !");
+                    }
+                    promotion.setAnnee(promotionDetails.getAnnee());
+                    return promotionRepository.save(promotion);
+                })
+                .orElseThrow(() -> new PromotionNotFoundException("Promotion avec ID " + id + " non trouvée"));
     }
+
 
     public void deletePromotion(Long id) {
+        if (!promotionRepository.existsById(id)) {
+            throw new PromotionNotFoundException("Promotion avec ID " + id + " non trouvée");
+        }
         promotionRepository.deleteById(id);
     }
-}
 
+    public boolean existsByAnnee(int annee) {
+        return promotionRepository.existsByAnnee(annee);
+    }
+}
